fix four HID + timer hardening bugs from the audit
* IHS_HIDManagerCloseAll iterated forward while
  IHS_HIDManagedDeviceClose shifted the array beneath it. With >1
  device the loop skipped elements, then crashed in IHS_MutexLock when
  IHS_ArrayListGet returned garbage memory. Reverse the loop like
  HIDPollTick does; drop the trailing no-op IHS_ArrayListClear.

* IHS_SessionChannelControlSendHIDMsg malloc'd a HID packet buffer and
  immediately packed into it without checking for allocator failure.
  Log + return false on NULL.

* IHS_HIDDeviceSDLGetFeatureReport had three issues in one function:
  (a) dereferenced reportNumber[0] without checking reportNumberLen,
  (b) SDL_GameControllerGetSerial can return NULL but strlen was
  called unconditionally,
  (c) the controllerType local was only assigned in the !xinput branch
  but read unconditionally below — uninitialized stack bytes flowed
  into the DeviceFeatureReport response sent to the host.

* IHS_TimerNow() computed tp.tv_sec * 1000 in time_t; on 32-bit
  time_t platforms this overflows at ~2.4M seconds of monotonic
  uptime. Cast to uint64_t before the multiply.

+++ src/ihs_timer.c
@@ -170,7 +170,9 @@ int IHS_TimerTaskGetRunCount(const IHS_TimerTask *task) {
 uint64_t IHS_TimerNow() {
     struct timespec tp;
     clock_gettime(CLOCK_MONOTONIC, &tp);
-    return tp.tv_sec * 1000 + tp.tv_nsec / 1000000;
+    // Cast to uint64_t before multiply: on platforms where time_t is 32-bit,
+    // `tp.tv_sec * 1000` overflows at ~2.4M seconds of monotonic uptime.
+    return (uint64_t) tp.tv_sec * 1000 + tp.tv_nsec / 1000000;
 }
 
 static void TimerThreadWorker() {
